+++ src/ApiClientCodeGen.VSIX/Options/GeneralOptionPage.cs
@@ -3,13 +3,6 @@ using Microsoft.VisualStudio.Shell;
 
 namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options
 {
-    public interface IGeneralOptions
-    {
-        string JavaPath { get; set; }
-        string NpmPath { get; set; }
-        string NSwagPath { get; set; }
-    }
-
     public class GeneralOptionPage : DialogPage, IGeneralOptions
     {
         public const string Name = "General";
