+++ src/ApiClientCodeGen.VSIX/Options/IGeneralOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options
+{
+    public interface IGeneralOptions
+    {
+        string JavaPath { get; set; }
+        string NpmPath { get; set; }
+        string NSwagPath { get; set; }
+    }
+}
