+++ src/ApiClientCodeGen.VSIX/Options/IOptionsFactory.cs
@@ -1,0 +1,8 @@
+﻿namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options
+{
+    public interface IOptionsFactory
+    {
+        TOptions Create<TOptions, TDialogPage>() 
+            where TOptions : class;
+    }
+}
