+++ src/ApiClientCodeGen.VSIX/Options/NSwagOptionsPage.cs
@@ -4,15 +4,6 @@ using NJsonSchema.CodeGeneration.CSharp;
 
 namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options
 {
-    public interface INSwagOption
-    {
-        bool InjectHttpClient { get; set; }
-        bool GenerateClientInterfaces { get; set; }
-        bool GenerateDtoTypes { get; set; }
-        bool UseBaseUrl { get; set; }
-        CSharpClassStyle ClassStyle { get; set; }
-    }
-
     public class NSwagOptionsPage : DialogPage, INSwagOption
     {
         public const string Name = "NSwag";
