+++ src/ApiClientCodeGen.VSIX/Options/OptionsFactory.cs
@@ -1,11 +1,5 @@
 ﻿namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options
 {
-    public interface IOptionsFactory
-    {
-        TOptions Create<TOptions, TDialogPage>() 
-            where TOptions : class;
-    }
-
     public class OptionsFactory : IOptionsFactory
     {
         public TOptions Create<TOptions, TDialogPage>()
